backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "logfire>=4.22.0",
    "pydantic>=2.12.5",
    "pydantic-ai>=1.55.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
